requirements.txt
pandas
numpy
psycopg2-binary
sqlalchemy
kafka-python
streamlit
scikit-learn
autots
jupyter